python_backend/requirements.txt
fastapi
uvicorn
transformers
torch
pymongo
python-dotenv
requests
